final push hairdresser and services

+++ packages/api/src/services/entities/service.entity.ts
@@ -1,4 +1,4 @@
-import { ObjectType, Field, Int, ID } from '@nestjs/graphql';
+import { ObjectType, Field, Int, ID, Float } from '@nestjs/graphql';
 import { Column, Entity, ObjectIdColumn } from 'typeorm';
 
 @Entity() // Database link - Typeorm
@@ -17,7 +17,7 @@ export class Service {
   description: string;
 
   @Column()
-  @Field(() => Int, { description: 'service price' })
+  @Field(() => Float, { description: 'service price' })
   price: number;
 
   @Column()
